evaluation/sugar-crepe/pyproject.toml
[project]
name = "screpe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py==2.3.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "albucore==0.0.24",
    "albumentations==2.0.8",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "array-record==0.8.0",
    "astunparse==1.6.3",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "boto3==1.40.14",
    "botocore==1.40.14",
    "braceexpand==0.1.7",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "clip>=0.2.0",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "dash>=3.2.0",
    "datasets==4.0.0",
    "dill==0.3.8",
    "dm-tree==0.1.9",
    "docstring-parser==0.17.0",
    "einops==0.8.1",
    "etils==1.13.0",
    "filelock==3.13.1",
    "flatbuffers==25.2.10",
    "fonttools==4.59.1",
    "frozenlist==1.7.0",
    "fsspec==2024.6.1",
    "ftfy==6.3.1",
    "future==1.0.0",
    "gast==0.6.0",
    "gdown==5.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-pasta==0.2.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "h5py==3.14.0",
    "hf-xet==1.1.8",
    "huggingface-hub==0.35.0rc0",
    "hydra-core==1.3.2",
    "idna==3.10",
    "imageio==2.37.0",
    "immutabledict==4.2.1",
    "importlib-resources==6.5.2",
    "jinja2==3.1.4",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "keras==3.11.2",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "libclang==18.1.1",
    "lightning-utilities==0.15.2",
    "loguru==0.7.3",
    "markdown==3.8.2",
    "markdown-it-py==4.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.9.0",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.3",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "multiprocess==0.70.16",
    "namex==0.1.0",
    "narwhals==2.6.0",
    "networkx==3.3",
    "nltk==3.9.1",
    "numpy==2.1.2",
    "omegaconf==2.3.0",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "opt-einsum==3.4.0",
    "optree==0.17.0",
    "packaging==25.0",
    "pandas==2.3.2",
    "pillow==11.0.0",
    "platformdirs==4.3.8",
    "plotly==6.3.0",
    "promise==2.3",
    "propcache==0.3.2",
    "protobuf==6.32.0",
    "psutil==7.0.0",
    "pyarrow==21.0.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2025.7.34",
    "requests==2.32.5",
    "rich==14.1.0",
    "s3transfer==0.13.1",
    "safetensors==0.6.2",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.1",
    "scipy==1.16.1",
    "seaborn==0.13.2",
    "sentry-sdk==2.35.0",
    "setproctitle==1.3.6",
    "setuptools==80.9.0",
    "simple-parsing==0.1.7",
    "simsimd==6.5.1",
    "six==1.17.0",
    "smmap==5.0.2",
    "soupsieve==2.7",
    "stringzilla==3.12.6",
    "sympy==1.13.1",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.20.0",
    "tensorflow-datasets==4.9.9",
    "tensorflow-metadata==1.17.2",
    "termcolor==3.1.0",
    "threadpoolctl==3.6.0",
    "tifffile==2025.10.4",
    "timm==1.0.19",
    "toml==0.10.2",
    "torch==2.6.0",
    "torch-tb-profiler==0.4.3",
    "torchmetrics==1.8.1",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "ujson==5.11.0",
    "umap-learn>=0.5.9.post2",
    "urllib3==2.5.0",
    "wandb==0.20.1",
    "wcwidth==0.2.13",
    "webdataset==1.0.2",
    "werkzeug==3.1.3",
    "wget==3.2",
    "wheel==0.45.1",
    "wordsegment==1.3.1",
    "wrapt==1.17.3",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
